Restore Linux support.

+++ src/dwire/DigiSpark.c
@@ -44,8 +44,8 @@ void WriteUPort(struct UPort *up) {
   Ws(", index ");     Wd(up->port.index,1);
   Ws(", character "); Wd(up->port.character,1);
   Ws(", baud ");      Wd(up->port.baud,1);
-  Ws(", device $");   Wx((int)up->device,1);
-  Ws(", handle $");   Wx((int)up->handle,1); Wsl(".");
+  Ws(", device $");   Wx((u64)up->device,1);
+  Ws(", handle $");   Wx((u64)up->handle,1); Wsl(".");
 }
